refactor(media): split stt orchestration and playback support
